requirements.txt
flask
mysql-connector-python
langchain
langgraph
pydantic
dotenv
